+++ packages/go_router/lib/src/builder.dart
@@ -49,6 +49,8 @@ class RouteBuilder {
   /// changes.
   final List<NavigatorObserver> observers;
 
+  final GoRouterStateRegistry _registry = GoRouterStateRegistry();
+
   /// Builds the top-level Navigator for the given [RouteMatchList].
   Widget build(
     BuildContext context,
@@ -61,17 +63,29 @@ class RouteBuilder {
       // empty box until then.
       return const SizedBox.shrink();
     }
-    try {
-      return tryBuild(
-          context, matchList, pop, routerNeglect, configuration.navigatorKey);
-    } on _RouteBuilderError catch (e) {
-      return _buildErrorNavigator(
-          context,
-          e,
-          Uri.parse(matchList.location.toString()),
-          pop,
-          configuration.navigatorKey);
-    }
+    return builderWithNav(
+      context,
+      Builder(
+        builder: (BuildContext context) {
+          try {
+            final Map<Page<Object?>, GoRouterState> newRegistry =
+                <Page<Object?>, GoRouterState>{};
+            final Widget result = tryBuild(context, matchList, pop,
+                routerNeglect, configuration.navigatorKey, newRegistry);
+            _registry.updateRegistry(newRegistry);
+            return GoRouterStateRegistryScope(
+                registry: _registry, child: result);
+          } on _RouteBuilderError catch (e) {
+            return _buildErrorNavigator(
+                context,
+                e,
+                Uri.parse(matchList.location.toString()),
+                pop,
+                configuration.navigatorKey);
+          }
+        },
+      ),
+    );
   }
 
   /// Builds the top-level Navigator by invoking the build method on each
@@ -85,21 +99,14 @@ class RouteBuilder {
     VoidCallback pop,
     bool routerNeglect,
     GlobalKey<NavigatorState> navigatorKey,
+    Map<Page<Object?>, GoRouterState> registry,
   ) {
     return builderWithNav(
       context,
-      GoRouterState(
-        configuration,
-        location: matchList.location.toString(),
-        name: null,
-        subloc: matchList.location.path,
-        queryParams: matchList.location.queryParameters,
-        queryParametersAll: matchList.location.queryParametersAll,
-        error: matchList.isError ? matchList.error : null,
-      ),
       _buildNavigator(
         pop,
-        buildPages(context, matchList, pop, routerNeglect, navigatorKey),
+        buildPages(
+            context, matchList, pop, routerNeglect, navigatorKey, registry),
         navigatorKey,
         observers: observers,
       ),
@@ -109,21 +116,22 @@ class RouteBuilder {
   /// Returns the top-level pages instead of the root navigator. Used for
   /// testing.
   @visibleForTesting
-  List<Page<dynamic>> buildPages(
+  List<Page<Object?>> buildPages(
       BuildContext context,
       RouteMatchList matchList,
       VoidCallback onPop,
       bool routerNeglect,
-      GlobalKey<NavigatorState> navigatorKey) {
+      GlobalKey<NavigatorState> navigatorKey,
+      Map<Page<Object?>, GoRouterState> registry) {
     try {
-      final Map<GlobalKey<NavigatorState>, List<Page<dynamic>>> keyToPage =
-          <GlobalKey<NavigatorState>, List<Page<dynamic>>>{};
+      final Map<GlobalKey<NavigatorState>, List<Page<Object?>>> keyToPage =
+          <GlobalKey<NavigatorState>, List<Page<Object?>>>{};
       final Map<String, String> params = <String, String>{};
       _buildRecursive(context, matchList, 0, onPop, routerNeglect, keyToPage,
-          params, navigatorKey);
+          params, navigatorKey, registry);
       return keyToPage[navigatorKey]!;
     } on _RouteBuilderError catch (e) {
-      return <Page<dynamic>>[
+      return <Page<Object?>>[
         _buildErrorPage(context, e, matchList.location),
       ];
     }
@@ -135,9 +143,10 @@ class RouteBuilder {
     int startIndex,
     VoidCallback pop,
     bool routerNeglect,
-    Map<GlobalKey<NavigatorState>, List<Page<dynamic>>> keyToPages,
+    Map<GlobalKey<NavigatorState>, List<Page<Object?>>> keyToPages,
     Map<String, String> params,
     GlobalKey<NavigatorState> navigatorKey,
+    Map<Page<Object?>, GoRouterState> registry,
   ) {
     if (startIndex >= matchList.matches.length) {
       return;
@@ -156,17 +165,17 @@ class RouteBuilder {
     };
     final GoRouterState state = buildState(match, newParams);
     if (route is GoRoute) {
-      final Page<dynamic> page = _buildPageForRoute(context, state, match);
-
+      final Page<Object?> page = _buildPageForRoute(context, state, match);
+      registry[page] = state;
       // If this GoRoute is for a different Navigator, add it to the
       // list of out of scope pages
       final GlobalKey<NavigatorState> goRouteNavKey =
           route.parentNavigatorKey ?? navigatorKey;
 
-      keyToPages.putIfAbsent(goRouteNavKey, () => <Page<dynamic>>[]).add(page);
+      keyToPages.putIfAbsent(goRouteNavKey, () => <Page<Object?>>[]).add(page);
 
       _buildRecursive(context, matchList, startIndex + 1, pop, routerNeglect,
-          keyToPages, newParams, navigatorKey);
+          keyToPages, newParams, navigatorKey, registry);
     } else if (route is ShellRouteBase) {
       assert(startIndex + 1 < matchList.matches.length,
           'Shell routes must always have child routes');
@@ -175,7 +184,7 @@ class RouteBuilder {
       final GlobalKey<NavigatorState> parentNavigatorKey = navigatorKey;
 
       // Add an entry for the parent navigator if none exists.
-      keyToPages.putIfAbsent(parentNavigatorKey, () => <Page<dynamic>>[]);
+      keyToPages.putIfAbsent(parentNavigatorKey, () => <Page<Object?>>[]);
 
       // Calling _buildRecursive can result in adding pages to the
       // parentNavigatorKey entry's list. Store the current length so
@@ -194,13 +203,13 @@ class RouteBuilder {
           'sub-routes');
 
       // Add an entry for the shell route's navigator
-      keyToPages.putIfAbsent(shellNavigatorKey!, () => <Page<dynamic>>[]);
+      keyToPages.putIfAbsent(shellNavigatorKey!, () => <Page<Object?>>[]);
 
-      // Build the remaining pages and retrieve the state for the top of the
-      // navigation stack
+      // Build the remaining pages
       _buildRecursive(context, matchList, startIndex + 1, pop, routerNeglect,
-          keyToPages, newParams, shellNavigatorKey);
+          keyToPages, newParams, shellNavigatorKey, registry);
 
+      // Build the Navigator and/or StatefulNavigationShell
       Widget child;
       if (route is StatefulShellRoute) {
         final String? restorationScopeId = route.branches
@@ -216,6 +225,7 @@ class RouteBuilder {
           navigator: child as Navigator,
           matchList: matchList,
           pop: pop,
+          registry: registry,
         );
       } else {
         final String? restorationScopeId =
@@ -226,19 +236,19 @@ class RouteBuilder {
       }
 
       // Build the Page for this route
-      final Page<dynamic> page =
+      final Page<Object?> page =
           _buildPageForRoute(context, state, match, child: child);
-
+      registry[page] = state;
       // Place the ShellRoute's Page onto the list for the parent navigator.
       keyToPages
-          .putIfAbsent(parentNavigatorKey, () => <Page<dynamic>>[])
+          .putIfAbsent(parentNavigatorKey, () => <Page<Object?>>[])
           .insert(shellPageIdx, page);
     }
   }
 
   Navigator _buildNavigator(
     VoidCallback pop,
-    List<Page<dynamic>> pages,
+    List<Page<Object?>> pages,
     Key? navigatorKey, {
     List<NavigatorObserver> observers = const <NavigatorObserver>[],
     String? restorationScopeId,
@@ -258,13 +268,13 @@ class RouteBuilder {
     );
   }
 
-  StatefulNavigationShell _buildStatefulNavigationShell({
-    required StatefulShellRoute shellRoute,
-    required Navigator navigator,
-    required GoRouterState shellRouterState,
-    required RouteMatchList matchList,
-    required VoidCallback pop,
-  }) {
+  StatefulNavigationShell _buildStatefulNavigationShell(
+      {required StatefulShellRoute shellRoute,
+      required Navigator navigator,
+      required GoRouterState shellRouterState,
+      required RouteMatchList matchList,
+      required VoidCallback pop,
+      required Map<Page<Object?>, GoRouterState> registry}) {
     return StatefulNavigationShell(
         configuration: configuration,
         shellRoute: shellRoute,
@@ -275,8 +285,8 @@ class RouteBuilder {
             RouteMatchList navigatorMatchList,
             GlobalKey<NavigatorState> navigatorKey,
             String? restorationScopeId) {
-          final List<Page<dynamic>> pages =
-              buildPages(context, navigatorMatchList, pop, true, navigatorKey);
+          final List<Page<dynamic>> pages = buildPages(
+              context, navigatorMatchList, pop, true, navigatorKey, registry);
           return _buildNavigator(pop, pages, navigatorKey,
               restorationScopeId: restorationScopeId);
         });
@@ -310,11 +320,11 @@ class RouteBuilder {
   }
 
   /// Builds a [Page] for [StackedRoute]
-  Page<dynamic> _buildPageForRoute(
+  Page<Object?> _buildPageForRoute(
       BuildContext context, GoRouterState state, RouteMatch match,
       {Widget? child}) {
     final RouteBase route = match.route;
-    Page<dynamic>? page;
+    Page<Object?>? page;
 
     if (route is GoRoute) {
       // Call the pageBuilder if it's non-null
@@ -342,8 +352,10 @@ class RouteBuilder {
 
     // Return the result of the route's builder() or pageBuilder()
     return page ??
-        buildPage(context, state,
-            _callRouteBuilder(context, state, match, childWidget: child));
+        // Uses a Builder to make sure its rebuild scope is limited to the page.
+        buildPage(context, state, Builder(builder: (BuildContext context) {
+          return _callRouteBuilder(context, state, match, childWidget: child);
+        }));
   }
 
   /// Calls the user-provided route builder from the [RouteMatch]'s [RouteBase].
@@ -427,7 +439,7 @@ class RouteBuilder {
 
   /// builds the page based on app type, i.e. MaterialApp vs. CupertinoApp
   @visibleForTesting
-  Page<dynamic> buildPage(
+  Page<Object?> buildPage(
     BuildContext context,
     GoRouterState state,
     Widget child,
@@ -464,7 +476,7 @@ class RouteBuilder {
       Uri uri, VoidCallback pop, GlobalKey<NavigatorState> navigatorKey) {
     return _buildNavigator(
       pop,
-      <Page<dynamic>>[
+      <Page<Object?>>[
         _buildErrorPage(context, e, uri),
       ],
       navigatorKey,